move actor_parser.py as it is format-agnostic and will be used for json, yaml, xml and rdf parsing
