+++ src/spdx/parser/actor_parser.py
